Refactor Trivia to DailyTrivia
