+++ client/src/components/DailyTrivia.tsx
